Layout.edje_get is better to return weak pointer otherwise one has to keep it forever (or put it in HashTable black hole)
NOTE: I have not changed Layout.edje_get type to Edje.Object which
should be best to do.

vapi/elm.vapi:
   - ListItem append, prepend, insert_before, insert_after: icon, end, and func may be null
   - add ListItem: selected_item_get and selected_items_get

NOTE: ListItem append etc could also return weak but for now I use
HashTable black hole to have the effect of weak (which was the problem of this thread).